AIoT/OpenVINO/model_server/example_client/object_detection/requirements.txt
grpcio==1.36.1
grpcio-tools==1.36.1
numpy
opencv-python
protobuf==3.15.6
tensorflow==2.*
tensorflow-serving-api==2.*
